push_swap

+++ push_swap.h
@@ -4,7 +4,7 @@
 #include <stdio.h>
 #include <unistd.h>
 #include <../libft/libft.h>
-
+#include <limits.h>
 typedef struct {
     char *long_number;
     char **numbers;
@@ -16,6 +16,6 @@ void free_double(char **tbl);
 int help_join(char *number);
 int ft_isvalid(char c);
 int ft_parsing(char **numbers);
-int *convert_to_int(char **numbers);
+long *convert_to_int(char **numbers);
 
 #endif
